Refactor import paths  and formatting
- Updated import path for AugmentationManager to reflect new module structure.
