+++ src/enrichment/themes_level_1.yaml
@@ -7,7 +7,8 @@
 07 - Infraestrutura e Transportes
 08 - Cultura, Artes e Patrimônio
 09 - Esportes e Lazer
-10 - Agricultura, Pecuária e Abastecimento11 - Indústria e Comércio
+10 - Agricultura, Pecuária e Abastecimento
+11 - Indústria e Comércio
 12 - Relações Internacionais e Diplomacia
 13 - Justiça e Direitos Humanos
 14 - Trabalho e Emprego
